+++ src/main/java/ru/philimonov/service/CreateAccountService.java
@@ -4,6 +4,8 @@ import ru.philimonov.converter.AccountToDtoConverter;
 import ru.philimonov.dao.AccountDao;
 import ru.philimonov.dao.AccountModel;
 
+import java.math.BigDecimal;
+
 public class CreateAccountService {
     private final AccountDao accountDao;
     private final AccountToDtoConverter accountToDtoConverter;
@@ -13,7 +15,7 @@ public class CreateAccountService {
         this.accountToDtoConverter = new AccountToDtoConverter();
     }
 
-    public AccountDto createAccount(String accountName, double amount, long id) {
+    public AccountDto createAccount(String accountName, BigDecimal amount, long id) {
         AccountModel accountModel = accountDao.createAccount(accountName, amount, id);
         if (accountModel == null) {
             return null;
